# internal/api/server.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Tier defines model for Tier.
type Tier struct {
	CreditsPerMonth int     `json:"creditsPerMonth"`
	Id              *string `json:"id,omitempty"`
	Name            string  `json:"name"`
}

// TierCreate defines model for TierCreate.
type TierCreate struct {
	CreditsPerMonth int    `json:"creditsPerMonth"`
	Name            string `json:"name"`
}

// CreateTierJSONRequestBody defines body for CreateTier for application/json ContentType.
type CreateTierJSONRequestBody = TierCreate

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create a new customer tier
	// (POST /admin/tier)
	CreateTier(w http.ResponseWriter, r *http.Request)
	// Enqueue usage-collection for a single app
	// (POST /admin/usage/app/{guid})
	CreateAppUsageJob(w http.ResponseWriter, r *http.Request, guid string)
	// Enqueue a full usage-collection job
	// (POST /admin/usage/job)
	CreateUsageJob(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Create a new customer tier
// (POST /admin/tier)
func (_ Unimplemented) CreateTier(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Enqueue usage-collection for a single app
// (POST /admin/usage/app/{guid})
func (_ Unimplemented) CreateAppUsageJob(w http.ResponseWriter, r *http.Request, guid string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Enqueue a full usage-collection job
// (POST /admin/usage/job)
func (_ Unimplemented) CreateUsageJob(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreateTier operation middleware
func (siw *ServerInterfaceWrapper) CreateTier(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{"usage.admin"})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTier(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateAppUsageJob operation middleware
func (siw *ServerInterfaceWrapper) CreateAppUsageJob(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "guid" -------------
	var guid string

	err = runtime.BindStyledParameterWithOptions("simple", "guid", chi.URLParam(r, "guid"), &guid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "guid", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{"usage.admin"})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateAppUsageJob(w, r, guid)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateUsageJob operation middleware
func (siw *ServerInterfaceWrapper) CreateUsageJob(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{"usage.admin"})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateUsageJob(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/admin/tier", wrapper.CreateTier)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/admin/usage/app/{guid}", wrapper.CreateAppUsageJob)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/admin/usage/job", wrapper.CreateUsageJob)
	})

	return r
}

type CreateTierRequestObject struct {
	Body *CreateTierJSONRequestBody
}

type CreateTierResponseObject interface {
	VisitCreateTierResponse(w http.ResponseWriter) error
}

type CreateTier201JSONResponse Tier

func (response CreateTier201JSONResponse) VisitCreateTierResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateAppUsageJobRequestObject struct {
	Guid string `json:"guid"`
}

type CreateAppUsageJobResponseObject interface {
	VisitCreateAppUsageJobResponse(w http.ResponseWriter) error
}

type CreateAppUsageJob202Response struct {
}

func (response CreateAppUsageJob202Response) VisitCreateAppUsageJobResponse(w http.ResponseWriter) error {
	w.WriteHeader(202)
	return nil
}

type CreateUsageJobRequestObject struct {
}

type CreateUsageJobResponseObject interface {
	VisitCreateUsageJobResponse(w http.ResponseWriter) error
}

type CreateUsageJob202Response struct {
}

func (response CreateUsageJob202Response) VisitCreateUsageJobResponse(w http.ResponseWriter) error {
	w.WriteHeader(202)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create a new customer tier
	// (POST /admin/tier)
	CreateTier(ctx context.Context, request CreateTierRequestObject) (CreateTierResponseObject, error)
	// Enqueue usage-collection for a single app
	// (POST /admin/usage/app/{guid})
	CreateAppUsageJob(ctx context.Context, request CreateAppUsageJobRequestObject) (CreateAppUsageJobResponseObject, error)
	// Enqueue a full usage-collection job
	// (POST /admin/usage/job)
	CreateUsageJob(ctx context.Context, request CreateUsageJobRequestObject) (CreateUsageJobResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// CreateTier operation middleware
func (sh *strictHandler) CreateTier(w http.ResponseWriter, r *http.Request) {
	var request CreateTierRequestObject

	var body CreateTierJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateTier(ctx, request.(CreateTierRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateTier")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateTierResponseObject); ok {
		if err := validResponse.VisitCreateTierResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateAppUsageJob operation middleware
func (sh *strictHandler) CreateAppUsageJob(w http.ResponseWriter, r *http.Request, guid string) {
	var request CreateAppUsageJobRequestObject

	request.Guid = guid

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateAppUsageJob(ctx, request.(CreateAppUsageJobRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateAppUsageJob")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateAppUsageJobResponseObject); ok {
		if err := validResponse.VisitCreateAppUsageJobResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateUsageJob operation middleware
func (sh *strictHandler) CreateUsageJob(w http.ResponseWriter, r *http.Request) {
	var request CreateUsageJobRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateUsageJob(ctx, request.(CreateUsageJobRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateUsageJob")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateUsageJobResponseObject); ok {
		if err := validResponse.VisitCreateUsageJobResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}